+++ projects/plugins/jetpack/_inc/client/recommendations/discount-badge/index.jsx
@@ -1,0 +1,32 @@
+/**
+ * External dependencies
+ */
+import { __, sprintf } from '@wordpress/i18n';
+import classNames from 'classnames';
+import React from 'react';
+
+/**
+ * Style dependencies
+ */
+import './style.scss';
+
+const DiscountBadge = ( { className, discount, suffix } ) => {
+	if ( ! discount ) {
+		return null;
+	}
+
+	return (
+		<div className={ classNames( 'jp-recommendations-discount-badge', className ) }>
+			{ /* eslint-disable */ }
+			{ sprintf(
+				// translators: %d is the percentage value, %% the percentage symbol
+				__( '%d%% off', 'jetpack' ), // @wordpress/valid-sprintf doesn't understand that the % symbol must be escaped
+				discount
+			) }
+			{ /* eslint-enable */ }
+			{ suffix }
+		</div>
+	);
+};
+
+export default DiscountBadge;