+++ projects/plugins/jetpack/_inc/client/recommendations/index.jsx
@@ -16,6 +16,7 @@ import { ProductPurchased } from './product-purchased';
 import { SiteTypeQuestion } from './prompts/site-type';
 import { Summary } from './summary';
 import QueryRecommendationsData from 'components/data/query-recommendations-data';
+import QueryIntroOffers from 'components/data/query-intro-offers';
 import QueryRecommendationsProductSuggestions from 'components/data/query-recommendations-product-suggestions';
 import QueryRecommendationsUpsell from 'components/data/query-recommendations-upsell';
 import QueryRecommendationsConditional from 'components/data/query-recommendations-conditional';
@@ -28,6 +29,7 @@ import {
 	isRecommendationsConditionalLoaded,
 } from 'state/recommendations';
 import { getNewRecommendations } from 'state/initial-state';
+import QuerySiteDiscount from 'components/data/query-site-discount';
 import { JetpackLoadingIcon } from 'components/jetpack-loading-icon';
 import { RECOMMENDATION_WIZARD_STEP } from './constants';
 import { ResourcePrompt } from './prompts/resource-prompt';
@@ -96,6 +98,8 @@ const RecommendationsComponent = props => {
 			<QueryRewindStatus />
 			<QuerySite />
 			<QuerySitePlugins />
+			<QuerySiteDiscount />
+			<QueryIntroOffers />
 			{ isLoading ? (
 				<div className="jp-recommendations__loading">
 					<JetpackLoadingIcon altText={ __( 'Loading recommendations', 'jetpack' ) } />